Nioclas Regnault: - Jack generator is now working
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1002 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnSphereShort.h
@@ -358,7 +358,7 @@ inline void BosonOnSphereShort::ConvertToMonomial(unsigned long* initialState, i
 {
   int Index = 0;
   for (int i = initialStateLzMax; i >= 0; --i)
-    for (int j = 0; j < initialState[i]; ++j)
+    for (unsigned long j = 0l; j < initialState[i]; ++j)
       finalState[Index++] = i;
 }
